Feat 배송 상세 조회

+++ delivery-service/src/main/java/com/firstlogistics/deliverservice/domain/event/OrderAcceptedEvent.java
@@ -39,6 +39,6 @@ public record OrderAcceptedEvent(
 		UUID productId,
 		String productName,
 		int quantity,
-		int price
+		Long price
 	) {}
 }
